Removed non-working combined run profile for the IDE
Created a simple grammar which has a lot of boilerplate and is able to parse number expressions: ()+-*/ and numbers with regex "[1-9]+".

For the grammar there is also a PrettyPrint.cs Visitor for console and CodeGen.cs for assigning each expression to an integer

input.mino is an example input program and output.ino is an example output Arduino program. Note: .ino files are in .gitignore, so it is required to run locally the program to create the output.ino file.